feat: Use a Toggle Switch to show external contributions
